feat: introduced support for DI container
In .net is common pratice to use a DI container to resolve the dependencies. This makes the application more testable and also a better UX for customers using a DI container.
It also makes it easier for users to extend the SDK by replacing an implementation on the DI container in a simple way.

To keep it easier for users not using a DI container 2 default constructors where added.